requirements.txt
cssmin==0.2.0
dj-database-url==0.4.2
django-assets==0.12
django-filter==1.0.4
django-jinja==2.3.1
djangorestframework==3.6.3
gunicorn==19.7.1
jsmin==2.2.2
Markdown==2.6.8
pipdeptree==0.10.1
psycopg2==2.7.1
pyparsing==2.2.0
python-dotenv==0.6.4
setuptools==36.2.0
six==1.10.0
wheel==0.29.0
whitenoise==3.3.0
